% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{from_imo}
\alias{from_imo}
\title{Uses an IMO number to obtain information about a vesel}
\usage{
from_imo(imo)
}
\arguments{
\item{imo}{International Maritime Organisation number}
}
\value{
Array of info describing the vessel
}
\description{
Uses an IMO number to obtain information about a vesel
}
