% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{from_name}
\alias{from_name}
\title{Uses a vessel's name to obtain information about a vesel}
\usage{
from_name(name)
}
\arguments{
\item{name}{The name of a vessel}
}
\value{
Array of info describing the vessel
}
\description{
Uses a vessel's name to obtain information about a vesel
}
